+++ android/samples/sample-gltf-viewer/src/main/java/com/google/android/filament/gltf/MainActivity.kt
@@ -21,20 +21,18 @@ import android.app.Activity
 import android.os.Bundle
 import android.util.Log
 import android.view.*
+import android.view.GestureDetector
 import android.widget.TextView
 import android.widget.Toast
-import com.google.android.filament.utils.KtxLoader
-import com.google.android.filament.utils.ModelViewer
-import com.google.android.filament.utils.RemoteServer
-import com.google.android.filament.utils.Utils
+import com.google.android.filament.utils.*
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
 import java.io.ByteArrayInputStream
-import java.io.IOException
 import java.nio.Buffer
 import java.nio.ByteBuffer
+import java.nio.charset.StandardCharsets
 import java.util.zip.ZipInputStream
 
 class MainActivity : Activity() {
@@ -55,6 +53,8 @@ class MainActivity : Activity() {
     private var remoteServer: RemoteServer? = null
     private var statusToast: Toast? = null
     private var statusText: String? = null
+    private var latestDownload: String? = null
+    private val automation = AutomationEngine()
 
     @SuppressLint("ClickableViewAccessibility")
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -223,6 +223,27 @@ class MainActivity : Activity() {
         choreographer.removeFrameCallback(frameScheduler)
     }
 
+    fun loadModelData(message: RemoteServer.ReceivedMessage) {
+        Log.i(TAG, "Downloaded model ${message.label} (${message.buffer.capacity()} bytes)")
+        clearStatusText()
+        titlebarHint.text = message.label
+        CoroutineScope(Dispatchers.IO).launch {
+            if (message.label.endsWith(".zip")) {
+                loadZip(message)
+            } else {
+                loadGlb(message)
+            }
+        }
+    }
+
+    fun loadSettings(message: RemoteServer.ReceivedMessage) {
+        val json = StandardCharsets.UTF_8.decode(message.buffer).toString()
+        automation.applySettings(json, modelViewer.view, null,
+                modelViewer.scene.indirectLight, modelViewer.light, modelViewer.engine.lightManager,
+                modelViewer.scene)
+        modelViewer.view.colorGrading = automation.getColorGrading((modelViewer.engine))
+    }
+
     inner class FrameCallback : Choreographer.FrameCallback {
         private val startTime = System.nanoTime()
         override fun doFrame(frameTimeNanos: Long) {
@@ -238,31 +259,25 @@ class MainActivity : Activity() {
 
             modelViewer.render(frameTimeNanos)
 
-            val label = remoteServer?.peekIncomingLabel()
-            if (label != null) {
-                Log.i(TAG, "Downloading $label")
-                setStatusText("Downloading $label")
+            // Check if a new download is in progress. If so, let the user know with toast.
+            val currentDownload = remoteServer?.peekIncomingLabel()
+            if (RemoteServer.isBinary(currentDownload) && currentDownload != latestDownload) {
+                latestDownload = currentDownload
+                Log.i(TAG, "Downloading $currentDownload")
+                setStatusText("Downloading $currentDownload")
             }
 
+            // Check if a new message has been fully received from the client.
             val message = remoteServer?.acquireReceivedMessage()
             if (message != null) {
-                Log.i(TAG, "Downloaded ${message.label} ${message.buffer.capacity()}")
-                clearStatusText()
-                titlebarHint.text = message.label
-
-                CoroutineScope(Dispatchers.IO).launch {
-                    try {
-                        if (message.label.endsWith(".zip")) {
-                            loadZip(message)
-                        } else {
-                            loadGlb(message)
-                        }
-                    } catch (exc: IOException) {
-                        setStatusText("URL fetch failed.")
-                        Log.e(TAG, "URL fetch failed", exc)
-                    }
+                if (message.label == latestDownload) {
+                    latestDownload = null
                 }
-
+                if (RemoteServer.isJson(message.label)) {
+                    loadSettings(message)
+                } else {
+                    loadModelData(message)
+                }
             }
         }
     }