+++ libs/filagui/include/filagui/ImGuiHelper.h
@@ -50,9 +50,8 @@ public:
     ImGuiHelper(filament::Engine* engine, filament::View* view, const utils::Path& fontPath);
     ~ImGuiHelper();
 
-    // Informs ImGui of the current display size, as well as the pixel ratio for high DPI displays.
-    // The display size is given in terms of virtual pixels, not physical pixels.
-    void setDisplaySize(int width, int height, float scaleX = 0.0f, float scaleY = 0.0f);
+    // Informs ImGui of the current display size, as well as a scaling factor when scissoring.
+    void setDisplaySize(int width, int height, float scaleX = 1.0f, float scaleY = 1.0f);
 
     // High-level utility method that takes a callback for creating all ImGui windows and widgets.
     // Clients are responsible for rendering the View. This should be called on every frame,
@@ -82,10 +81,12 @@ public:
       filament::View* mView; // The view is owned by the client.
       filament::Scene* mScene;
       filament::Material* mMaterial = nullptr;
+      filament::Camera* mCamera = nullptr;
       std::vector<filament::VertexBuffer*> mVertexBuffers;
       std::vector<filament::IndexBuffer*> mIndexBuffers;
       std::vector<filament::MaterialInstance*> mMaterialInstances;
       utils::Entity mRenderable;
+      utils::Entity mCameraEntity;
       filament::Texture* mTexture = nullptr;
       bool mHasSynced = false;
       ImGuiContext* mImGuiContext;