+++ libs/viewer/include/viewer/Settings.h
@@ -18,7 +18,10 @@
 #define VIEWER_SETTINGS_H
 
 #include <filament/ColorGrading.h>
+#include <filament/IndirectLight.h>
+#include <filament/LightManager.h>
 #include <filament/MaterialInstance.h>
+#include <filament/Scene.h>
 #include <filament/View.h>
 
 #include <math/vec3.h>
@@ -37,6 +40,7 @@ struct DynamicLightingSettings;
 struct MaterialSettings;
 struct Settings;
 struct ViewSettings;
+struct LightSettings;
 
 using AmbientOcclusionOptions = filament::View::AmbientOcclusionOptions;
 using AntiAliasing = filament::View::AntiAliasing;
@@ -50,33 +54,35 @@ using TemporalAntiAliasingOptions = filament::View::TemporalAntiAliasingOptions;
 using ToneMapping = filament::ColorGrading::ToneMapping;
 using VignetteOptions = filament::View::VignetteOptions;
 using VsmShadowOptions = filament::View::VsmShadowOptions;
-
-// Reads the given JSON blob and updates the corresponding fields in the given Settings object.
-// - The given JSON blob need not specify all settings.
-// - Returns true if successful.
-// - This function writes warnings and error messages into the utils log.
-bool readJson(const char* jsonChunk, size_t size, Settings* out);
+using LightManager = filament::LightManager;
 
 // These functions push all editable property values to their respective Filament objects.
 void applySettings(const ViewSettings& settings, View* dest);
 void applySettings(const MaterialSettings& settings, MaterialInstance* dest);
+void applySettings(const LightSettings& settings, IndirectLight* ibl, utils::Entity sunlight,
+        LightManager* lm, Scene* scene);
 
 // Creates a new ColorGrading object based on the given settings.
 ColorGrading* createColorGrading(const ColorGradingSettings& settings, Engine* engine);
 
-// Generates human-readable JSON strings from settings objects.
-std::string writeJson(const AmbientOcclusionOptions& in);
-std::string writeJson(const BloomOptions& in);
-std::string writeJson(const ColorGradingSettings& in);
-std::string writeJson(const DepthOfFieldOptions& in);
-std::string writeJson(const DynamicLightingSettings& in);
-std::string writeJson(const FogOptions& in);
-std::string writeJson(const MaterialSettings& in);
-std::string writeJson(const RenderQuality& in);
-std::string writeJson(const Settings& in);
-std::string writeJson(const TemporalAntiAliasingOptions& in);
-std::string writeJson(const ViewSettings& in);
-std::string writeJson(const VignetteOptions& in);
+class JsonSerializer {
+public:
+    JsonSerializer();
+    ~JsonSerializer();
+
+    // Writes a human-readable JSON string into an internal buffer and returns the result.
+    const std::string& writeJson(const Settings& in);
+
+    // Reads the given JSON blob and updates the corresponding fields in the given Settings object.
+    // - The given JSON blob need not specify all settings.
+    // - Returns true if successful.
+    // - This function writes warnings and error messages into the utils log.
+    bool readJson(const char* jsonChunk, size_t size, Settings* out);
+
+private:
+    class Context;
+    Context* context;
+};
 
 struct ColorGradingSettings {
     bool enabled = true;
@@ -140,9 +146,21 @@ struct MaterialSettings {
     MaterialProperty<math::float4> float4[MAX_COUNT];
 };
 
+struct LightSettings {
+    bool enableShadows = true;
+    bool enableSunlight = true;
+    LightManager::ShadowOptions shadowOptions;
+    float sunlightIntensity = 100000.0f;
+    math::float3 sunlightDirection = {0.6, -1.0, -0.8};;
+    math::float3 sunlightColor = filament::Color::toLinear<filament::ACCURATE>({ 0.98, 0.92, 0.89});
+    float iblIntensity = 30000.0f;
+    float iblRotation = 0.0f;
+};
+
 struct Settings {
     ViewSettings view;
     MaterialSettings material;
+    LightSettings lighting;
 };
 
 } // namespace viewer