+++ libs/viewer/include/viewer/SimpleViewer.h
@@ -33,6 +33,10 @@
 
 #include <math/vec3.h>
 
+namespace filagui {
+    class ImGuiHelper;
+}
+
 namespace filament {
 namespace viewer {
 
@@ -109,6 +113,19 @@ public:
     void updateUserInterface();
 
     /**
+     * Alternative to updateUserInterface that uses an internal instance of ImGuiHelper.
+     *
+     * This utility method is designed for clients that do not want to manage their own instance of
+     * ImGuiHelper (e.g., JavaScript clients).
+     *
+     * Behind the scenes this simply calls ImGuiHelper->render() and passes updateUserInterface into
+     * its callback. Note that the first call might be slower since it requires the creation of the
+     * internal ImGuiHelper instance.
+     */
+    void renderUserInterface(float timeStepInSeconds, filament::View* guiView, float pixelRatio,
+            float mouseX, float mouseY, bool mouseButton, float mouseWheelY);
+
+    /**
      * Retrieves the current width of the ImGui "window" which we are using as a sidebar.
      * Clients can monitor this value to adjust the size of the view.
      */
@@ -129,7 +146,7 @@ public:
      * Enables a built-in light source (useful for creating shadows).
      * Defaults to true.
      */
-    void enableSunlight(bool b) { mEnableSunlight = b; }
+    void enableSunlight(bool b) { mSettings.lighting.enableSunlight = b; }
 
     /**
      * Enables dithering on the view.
@@ -170,7 +187,7 @@ public:
      * See also filament::IndirectLight::setIntensity().
      * Defaults to 30000.0.
      */
-    void setIBLIntensity(float brightness) { mIblIntensity = brightness; }
+    void setIBLIntensity(float brightness) { mSettings.lighting.iblIntensity = brightness; }
 
     /**
      * Gets a modifiable reference to stashed state.
@@ -188,6 +205,9 @@ private:
     filament::View* const mView;
     const utils::Entity mSunlight;
 
+    // Lazily instantiated fields.
+    filagui::ImGuiHelper* mImGuiHelper = nullptr;
+
     // Properties that can be changed from the application.
     FilamentAsset* mAsset = nullptr;
     Animator* mAnimator = nullptr;
@@ -197,21 +217,15 @@ private:
     // Properties that can be changed from the UI.
     int mCurrentAnimation = 1;
     bool mResetAnimation = true;
-    float mIblIntensity = 30000.0f;
-    float mIblRotation = 0.0f;
-    float mSunlightIntensity = 100000.0f; // <-- This value is overridden when loading an IBL.
-    filament::math::float3 mSunlightColor = filament::Color::toLinear<filament::ACCURATE>({ 0.98, 0.92, 0.89});
-    filament::math::float3 mSunlightDirection = {0.6, -1.0, -0.8};
     bool mEnableWireframe = false;
-    bool mEnableSunlight = true;
-    bool mEnableShadows = true;
-    int mShadowCascades = 1;
-    bool mEnableContactShadows = false;
     int mVsmMsaaSamplesLog2 = 1;
-    std::array<float, 3> mSplitPositions = {0.25f, 0.50f, 0.75f};
     Settings mSettings;
     int mSidebarWidth;
     uint32_t mFlags;
+
+    // Color grading UI state.
+    float mRangePlot[1024 * 3];
+    float mCurvePlot[1024 * 3];
 };
 
 filament::math::mat4f fitIntoUnitCube(const filament::Aabb& bounds);